[Lombok] Extract `ToStringGenerator` out of `kotlin` subpackage because now it also operates on Java classes as well
