+++ plugins/lombok/lombok.k2/src/org/jetbrains/kotlin/lombok/k2/FirLombokCommonRegistrar.kt
@@ -9,12 +9,14 @@ import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrar
 import org.jetbrains.kotlin.lombok.k2.config.LombokService
 import org.jetbrains.kotlin.lombok.k2.generators.LoggerGenerator
 import org.jetbrains.kotlin.lombok.k2.generators.LombokConstructorsGenerator
+import org.jetbrains.kotlin.lombok.k2.generators.ToStringGenerator
 import java.io.File
 
 class FirLombokCommonRegistrar(private val lombokConfigFile: File?) : FirExtensionRegistrar() {
     override fun ExtensionRegistrarContext.configurePlugin() {
         +LombokService.getFactory(lombokConfigFile)
         +::LoggerGenerator
+        +::ToStringGenerator
         +::LombokConstructorsGenerator
 
         registerDiagnosticContainers(LombokCliDiagnostics)
