+++ plugins/lombok/lombok.k2/src/org/jetbrains/kotlin/lombok/k2/generators/kotlin/ir/ToStringIrGenerator.kt
@@ -27,7 +27,7 @@ import org.jetbrains.kotlin.ir.util.functions
 import org.jetbrains.kotlin.ir.util.isInterface
 import org.jetbrains.kotlin.ir.visitors.IrVisitorVoid
 import org.jetbrains.kotlin.ir.visitors.acceptChildrenVoid
-import org.jetbrains.kotlin.lombok.k2.generators.kotlin.ToStringGeneratorKey
+import org.jetbrains.kotlin.lombok.k2.generators.ToStringGeneratorKey
 import org.jetbrains.kotlin.util.OperatorNameConventions
 import org.jetbrains.kotlin.utils.addToStdlib.runIf
